+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ UNRELEASED
   * CHANGED:   Define ADAT_RX renamed to XUA_ADAT_RX_EN
   * CHANGED:   Define ADAT_TX renamed to XUA_ADAT_TX_EN
   * CHANGED:   Define SPDIF_RX renamed to XUA_SPDIF_RX_EN
+  * CHANGED:   Define SELF_POWERED changed to XUA_POWERMODE and associated
+    defines
   * CHANGED:   Drive strength of I2S clock lines upped to 8mA on xCORE.ai
   * CHANGED:   ADC datalines sampled on falling edge of clock in TDM mode
   * CHANGED:   Improved startup behaviour of TDM clocks